+++ src/migrations/Install.php
@@ -1,0 +1,37 @@
+<?php
+
+namespace digitalpulsebe\craftmultitranslator\migrations;
+
+use Craft;
+use craft\db\Migration;
+class Install extends Migration
+{
+    /**
+     * @inheritdoc
+     */
+    public function safeUp(): bool
+    {
+        $this->createTable('{{%multitranslator_deepl_glossaries}}', [
+            'id' => $this->primaryKey(),
+            'name' => $this->string(),
+            'deeplId' => $this->string(),
+            'sourceLanguage' => $this->string(5),
+            'targetLanguage' => $this->string(5),
+            'data' => $this->json(),
+            'dateCreated' => $this->dateTime()->notNull(),
+            'dateUpdated' => $this->dateTime()->notNull(),
+            'uid' => $this->uid(),
+        ]);
+
+        return true;
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function safeDown(): bool
+    {
+        $this->dropTableIfExists('{{%multitranslator_deepl_glossaries}}');
+        return true;
+    }
+}